1. use uint64_t for timestamp so that it works for 32 and 64 bit binaries; 2. fix: closes #272 and addresses #238; 3. remove: test-stensal-timestamp.c

+++ orka-utils.h
@@ -40,8 +40,8 @@ void orka_settings_init (struct orka_settings*, const char filename[]);
 int orka_iso8601_to_unix_ms(char *timestamp, size_t len, void *p_data);
 int orka_strtoull(char *str, size_t len, void *p_data) ;
 
-void orka_sleep_ms(const long long delay_ms);
-int64_t orka_timestamp_ms();
+void orka_sleep_ms(const int64_t delay_ms);
+uint64_t orka_timestamp_ms();
 void orka_timestamp_str(char *p_str, int len);
 
 int orka_str_to_ntl(
